requirements_web_demo.txt
# Core dependencies
gradio==5.49.1
gradio_client==1.13.3
transformers-stream-generator==0.0.5
git+https://github.com/huggingface/transformers.git@7aa888b7fa477d13153ffbfe107dfbd6c696014a
torch
torchvision
accelerate

# Optional dependency
# Uncomment the following line if you need flash-attn
flash-attn